increase GenericButton::Counters members type to int32_t
GenericButton::Counters::repeat size does not match with EventMsg::cntr, it resulted
in AutoRepeat event reporting wrong conunt number for repetitions >127

+++ src/espasyncbutton.hpp
@@ -200,8 +200,8 @@ protected:
 
   struct Counters
   {
-    int8_t click{0};        // click counter
-    int8_t repeat{0};       // repeat counter
+    int32_t click{0};        // click counter
+    int32_t repeat{0};       // repeat counter
   };
 
   /**
